Split emitted signals from the student models file.
