+++ common/djangoapps/student/signals/__init__.py
@@ -1,0 +1,6 @@
+from student.signals.signals import (
+    ENROLLMENT_TRACK_UPDATED,
+    UNENROLL_DONE,
+    ENROLL_STATUS_CHANGE,
+    REFUND_ORDER
+)
