+++ lms/djangoapps/commerce/signals.py
@@ -19,7 +19,7 @@ from openedx.core.djangoapps.commerce.utils import ecommerce_api_client, is_comm
 from openedx.core.djangoapps.site_configuration import helpers as configuration_helpers
 from openedx.core.djangoapps.theming import helpers as theming_helpers
 from request_cache.middleware import RequestCache
-from student.models import REFUND_ORDER
+from student.signals import REFUND_ORDER
 
 log = logging.getLogger(__name__)
