+++ lms/djangoapps/grades/signals/handlers.py
@@ -9,7 +9,7 @@ from django.dispatch import receiver
 from openedx.core.djangoapps.course_groups.signals.signals import COHORT_MEMBERSHIP_UPDATED
 from openedx.core.lib.grade_utils import is_score_higher_or_equal
 from student.models import user_by_anonymous_id
-from student.signals.signals import ENROLLMENT_TRACK_UPDATED
+from student.signals import ENROLLMENT_TRACK_UPDATED
 from submissions.models import score_reset, score_set
 from track.event_transaction_utils import get_event_transaction_id, get_event_transaction_type
 from util.date_utils import to_timestamp
